@@ -39,34 +39,30 @@
     "generate:docker-compose": "ts-node tools/generators/docker-compose.ts"
   },
   "dependencies": {
-    "@angular/animations": "13.1.1",
-    "@angular/common": "13.1.1",
-    "@angular/compiler": "13.1.1",
-    "@angular/core": "13.1.1",
-    "@angular/forms": "13.1.1",
-    "@angular/platform-browser": "13.1.1",
-    "@angular/platform-browser-dynamic": "13.1.1",
-    "@angular/platform-server": "13.1.1",
-    "@angular/router": "13.1.1",
-    "@capacitor/core": "^2.4.5",
-    "@ionic-native/core": "^5.36.0",
-    "@ionic-native/splash-screen": "^5.36.0",
-    "@ionic-native/status-bar": "^5.36.0",
+    "@angular/animations": "15.0.3",
+    "@angular/common": "15.0.3",
+    "@angular/compiler": "15.0.3",
+    "@angular/core": "15.0.3",
+    "@angular/forms": "15.0.3",
+    "@angular/platform-browser": "15.0.3",
+    "@angular/platform-browser-dynamic": "15.0.3",
+    "@angular/platform-server": "15.0.3",
+    "@angular/router": "15.0.3",
     "@ionic/angular": "^6.4.0",
     "@ionic/angular-server": "^6.4.0",
-    "@nestjs/common": "^8.0.0",
-    "@nestjs/core": "^8.0.0",
-    "@nestjs/platform-express": "^8.4.0",
+    "@nestjs/common": "9.0.0",
+    "@nestjs/core": "9.0.0",
+    "@nestjs/platform-express": "9.0.0",
     "@ngneat/until-destroy": "^9.2.2",
-    "@nguniversal/express-engine": "^12.1.0",
-    "@nrwl/angular": "13.4.3",
+    "@nguniversal/express-engine": "15.0.0",
+    "@nrwl/angular": "15.3.3",
     "express": "^4.15.2",
     "handlebars": "^4.7.7",
     "neo4j-driver": "^4.3.3",
     "parse": "^3.5.1",
     "parse-server": "^5.4.0",
     "reflect-metadata": "^0.1.13",
-    "rxjs": "^7.0.0",
+    "rxjs": "~7.5.0",
     "simple-parse-smtp-adapter": "github:music-bat/simple-parse-smtp-adapter",
     "spotify-web-api-node": "^5.0.2",
     "swiper": "^8.4.5",
@@ -74,54 +70,50 @@
     "zone.js": "0.11.4"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^13.0.0",
-    "@angular-devkit/build-angular": "13.3.9",
-    "@angular-eslint/eslint-plugin": "13.0.1",
-    "@angular-eslint/eslint-plugin-template": "13.0.1",
-    "@angular-eslint/template-parser": "13.0.1",
-    "@angular/cli": "13.1.2",
-    "@angular/compiler-cli": "13.1.1",
-    "@angular/language-service": "13.1.1",
-    "@capacitor/android": "^2.4.5",
-    "@capacitor/cli": "^3.3.3",
-    "@capacitor/ios": "^2.4.5",
+    "@angular-builders/custom-webpack": "^15.0.0-beta.0",
+    "@angular-devkit/build-angular": "15.0.3",
+    "@angular-eslint/eslint-plugin": "15.0.0",
+    "@angular-eslint/eslint-plugin-template": "15.0.0",
+    "@angular-eslint/template-parser": "15.0.0",
+    "@angular/cli": "~15.0.0",
+    "@angular/compiler-cli": "15.0.3",
+    "@angular/language-service": "15.0.3",
     "@ionic/angular-toolkit": "^7.0.0",
-    "@nestjs/schematics": "^8.0.0",
-    "@nestjs/testing": "^8.0.0",
-    "@nguniversal/builders": "^12.1.0",
-    "@nrwl/cli": "13.4.3",
-    "@nrwl/cypress": "13.4.3",
-    "@nrwl/eslint-plugin-nx": "13.4.3",
-    "@nrwl/jest": "13.4.3",
-    "@nrwl/linter": "13.4.3",
-    "@nrwl/nest": "13.4.3",
-    "@nrwl/node": "13.4.3",
-    "@nrwl/tao": "13.4.3",
-    "@nrwl/workspace": "13.4.3",
-    "@nxtend/capacitor": "^11.0.0",
-    "@nxtend/ionic-angular": "^13.1.0",
+    "@nestjs/schematics": "9.0.0",
+    "@nestjs/testing": "9.0.0",
+    "@nguniversal/builders": "15.0.0",
+    "@nrwl/cli": "15.3.3",
+    "@nrwl/cypress": "15.3.3",
+    "@nrwl/eslint-plugin-nx": "15.3.3",
+    "@nrwl/jest": "15.3.3",
+    "@nrwl/linter": "15.3.3",
+    "@nrwl/nest": "15.3.3",
+    "@nrwl/node": "15.3.3",
+    "@nrwl/workspace": "15.3.3",
     "@types/express": "^4.17.0",
     "@types/generate-password": "^1.5.1",
-    "@types/jest": "27.0.2",
+    "@types/jest": "28.1.8",
     "@types/js-yaml": "^4.0.3",
-    "@types/node": "^14.14.33",
+    "@types/node": "18.7.1",
     "@types/parse": "^3.0.1",
-    "@typescript-eslint/eslint-plugin": "5.3.1",
-    "@typescript-eslint/parser": "5.3.1",
+    "@typescript-eslint/eslint-plugin": "5.36.1",
+    "@typescript-eslint/parser": "5.36.1",
     "all-contributors-cli": "^6.20.0",
-    "cypress": "^7.7.0",
+    "cypress": "^11.2.0",
     "dotenv": "^10.0.0",
-    "eslint": "8.2.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-cypress": "^2.10.3",
+    "eslint-plugin-cypress": "^2.12.1",
     "generate-password": "^1.6.1",
-    "jest": "27.2.3",
-    "jest-preset-angular": "11.0.0",
+    "jest": "28.1.3",
+    "jest-environment-jsdom": "28.1.1",
+    "jest-preset-angular": "12.2.3",
     "js-yaml": "^4.1.0",
-    "prettier": "2.8.1",
-    "ts-jest": "27.0.5",
-    "ts-node": "^9.1.1",
-    "typescript": "4.5.4",
+    "nx": "15.3.3",
+    "prettier": "^2.8.1",
+    "ts-jest": "28.0.8",
+    "ts-node": "10.9.1",
+    "typescript": "4.8.4",
     "webpack-dev-server": "^4.7.2"
   }
 }
